backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
pydantic==2.10.6
